refactor(ts-migration): move runner to typescript
